+++ trailblazer/services/analysis_service/utils.py
@@ -1,4 +1,6 @@
 from trailblazer.constants import TrailblazerStatus
+from trailblazer.dto.analyses_response import UpdateAnalysesResponse
+from trailblazer.dto.analysis_response import AnalysisResponse
 from trailblazer.dto.summaries_response import Summary
 from trailblazer.store.models import Analysis
 
@@ -21,3 +23,18 @@ def create_summary(analyses: list[Analysis]) -> Summary:
         cancelled=cancelled,
         failed=failed,
     )
+
+def create_analysis_response(analysis: Analysis) -> AnalysisResponse:
+    analysis_data: dict = analysis.to_dict()
+    analysis_data["jobs"] = [job.to_dict() for job in analysis.analysis_jobs]
+    analysis_data["upload_jobs"] = [job.to_dict() for job in analysis.upload_jobs]
+    analysis_data["user"] = analysis.user.to_dict() if analysis.user else None
+    return AnalysisResponse.model_validate(analysis_data)
+
+
+def create_update_analyses_response(analyses: list[Analysis]) -> UpdateAnalysesResponse:
+    response_data: list[dict] = []
+    for analysis in analyses:
+        analysis_data = analysis.to_dict()
+        response_data.append(analysis_data)
+    return UpdateAnalysesResponse(analyses=response_data)
