+++ trailblazer/services/job_service/utils.py
@@ -1,3 +1,4 @@
+from trailblazer.dto.failed_jobs_response import FailedJobsResponse
 from trailblazer.dto.job_response import JobResponse
 from trailblazer.store.models import Job
 
@@ -6,3 +7,7 @@ def create_job_response(job: Job) -> JobResponse:
     return JobResponse(
         slurm_id=job.slurm_id, analysis_id=job.analysis_id, status=job.status, id=job.id
     )
+
+
+def create_failed_jobs_response(failed_job_statistics: list[dict]) -> FailedJobsResponse:
+    return FailedJobsResponse(jobs=failed_job_statistics)
